Rename DTT0Calibration plugins
Renamed DTT0Calibration to DTT0CalibrationRMS as it represents a
simple baseline based on the peak RMS.
Accordingly the more sophisticated DTT0CalibrationNew plugin is
now renamed to DTT0Calibration
