+++ CalibMuon/DTCalibration/plugins/DTT0CalibrationRMS.h
@@ -1,14 +1,12 @@
-#ifndef CalibMuon_DTT0CalibrationNew_H
-#define CalibMuon_DTT0CalibrationNew_H
+#ifndef CalibMuon_DTT0CalibrationRMS_H
+#define CalibMuon_DTT0CalibrationRMS_H
 
-/** \class DTT0CalibrationNew
+/** \class DTT0CalibrationRMS
  *  Analyzer class computes the mean and RMS of t0 from pulses.
  *  Those values are written in the DB with cell granularity. The
  *  mean value for each channel is normalized to a reference time common to all the sector.
- *  The t0 of wires in odd layers are corrected for the relative difference between 
- *  odd and even layers 
- *
- *  \author S. Bolognesi - INFN Torino
+ *  The t0 of wires in odd layers are corrected for the relative difference between
+ *  odd and even layers
  */
 
 #include "FWCore/Framework/interface/EDAnalyzer.h"
@@ -23,24 +21,23 @@
 class TFile;
 class TH1I;
 class TH1D;
-class TSpectrum;
 class DTT0;
 
-class DTT0CalibrationNew : public edm::EDAnalyzer {
+class DTT0CalibrationRMS : public edm::EDAnalyzer {
 public:
   /// Constructor
-  DTT0CalibrationNew(const edm::ParameterSet& pset);
+  DTT0CalibrationRMS(const edm::ParameterSet& pset);
 
   /// Destructor
-  virtual ~DTT0CalibrationNew();
+  ~DTT0CalibrationRMS() override;
 
   // Operations
 
   /// Fill the maps with t0 (by channel)
-  void analyze(const edm::Event & event, const edm::EventSetup& eventSetup);
+  void analyze(const edm::Event & event, const edm::EventSetup& eventSetup) override;
 
   /// Compute the mean and the RMS of the t0 from the maps and write them to the DB with channel granularity
-  void endJob();
+  void endJob() override;
 
 
 protected:
@@ -56,8 +53,6 @@ private:
   // The label used to retrieve digis from the event
   std::string digiLabel;
 
-  std::string dbLabel;
-
   // The root file which contain the histos per layer
   TFile *theFile;
   // The root file which will contain the histos per wire (for the given layer)
@@ -70,19 +65,14 @@ private:
   //Number of events to be used for the t0 reference per wire
   unsigned int eventsForWireT0;
 
+  //Reject digis if far from TP peak
+  unsigned int rejectDigiFromPeak;
+
   //Acceptance of t0 w.r.t. reference peak
   double tpPeakWidth;
 
-  //Acceptance of t0 w.r.t. reference peak
-  double tpPeakWidthPerLayer;
-
-  //Size of timeBox 
-  unsigned int timeBoxWidth; 
-
-  //Digi's will be rejected if too far from TP peak 
-  unsigned int rejectDigiFromPeak;
-
-  unsigned int retryForLayerT0;
+  // Write relative t0's with respect to mean t0's in chamber
+  bool correctByChamberMean_;
 
   //The wheels,sector to be calibrated (default All)
   std::string theCalibWheel;
@@ -95,8 +85,6 @@ private:
   //Histo with t0 mean per layer for all the sector
   TH1D* hT0SectorHisto;
   
-  TSpectrum* spectrum;
-
   //Layer with histos for each wire
   std::vector<DTWireId> wireIdWithHistos;
   std::vector<std::string> cellsWithHistos;
@@ -112,17 +100,12 @@ private:
   std::map<DTWireId,double> qK;
   //Map with histo per wire for the chosen layer
   std::map<DTWireId,TH1I*> theHistoWireMap;
+  std::map<DTWireId,TH1I*> theHistoWireMap_ref;
   //Map with mean and RMS of t0 per layer
   std::map<std::string,double> theT0LayerMap;
   std::map<std::string,double> theSigmaT0LayerMap;
-  std::map<DTLayerId,double> theTPPeakMap;
-  //Ref. t0 per chamber
-  std::map<DTChamberId,double> theSumT0ByChamber;
-  std::map<DTChamberId,int> theCountT0ByChamber;
-  std::map<DTChamberId,double> theRefT0ByChamber;
 
   //DTGeometry used to loop on the SL in the endJob
   edm::ESHandle<DTGeometry> dtGeom;
 };
 #endif
-
